pyproject.toml
[project]
name = "devis-app"
version = "0.1.0"
description = "Application Flask de gestion de devis et factures"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.1",
    "flask-mail>=0.9.1",
    "weasyprint>=60.0",
    "wtforms>=3.1.1",
    "xhtml2pdf>=0.2.17",
    "flask-login>=0.6.3",
    "python-dotenv>=1.2.1",
    "gunicorn>=25.0.1",
    "psycopg2-binary>=2.9.11",
    "email-validator>=2.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
